+++ src/components/export-to-csv/index.tsx
@@ -35,7 +35,7 @@ export default function ExportToCSV({
       {!doneExporting ? (
         <button
           onClick={() => handleDownload()}
-          className="w-fit relative py-2.5 px-6 bg-[#F2F4F7] text-[#1D2939] hover:bg-primary transition-all hover:text-white rounded-full flex items-center gap-3"
+          className="w-fit text-nowrap relative py-2.5 px-6 bg-[#F2F4F7] text-[#1D2939] hover:bg-primary transition-all hover:text-white rounded-full flex items-center gap-3"
         >
           {isExporting ? "...loading" : "Export to CSV"}
         </button>