+++ src/pages/bookings/requests/index.tsx
@@ -48,7 +48,7 @@ export default function Requests() {
           <div className=" w-full flex justify-between flex-col md:flex-row gap-3">
             <h2 className="text-xl font-semibold">All Requests List</h2>
             <div className=" flex items-center gap-4 flex-col md:flex-row ">
-              <ExportSelect id="bookings" />
+              {/* <ExportSelect id="bookings" /> */}
               <LoadingButton
                 clickHandler={() => setOpenNewRequest(true)}
                 isLoading={false}